pyproject.toml
[tool.poetry]
name = "dbx-deploy"
version = "0.13.2"
description = "Databricks Deployment Tool"
readme = "README.md"
repository = "https://github.com/daipe-ai/dbx-deploy"

license = "MIT"

authors = [
    "Jiri Koutny <jiri.koutny@datasentics.com>"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

include = ["*.yaml", "*.tpl"]

packages = [
    { include = "dbxdeploy", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
databricks-api = ">=0.3.0,<1.0.0"
dbx-notebook-exporter = "0.4.*"
nbconvert = "^5.6"
pygit2 = "^1.3"
python-box = "^3.4"
tomlkit = ">=0.5.8,<1.0.0"
pyfony-bundles = "0.4.*"
pyfony-core = "0.8.*"
console-bundle = "0.4.*"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.0"
flake8-fixme = "^1.1.1"
pep8-naming = "^0.11.1"
poethepoet = "^0.10.0"
black = "^20.8b1"

[tool.poetry.plugins."pyfony.bundle"]
create = "dbxdeploy.DbxDeployBundle:DbxDeployBundle"

[pyfony.bootstrap]
container_init = "pyfonycore.container.container_init:init_with_current_bundle"
root_module_name = "dbxdeploy"

[tool.black]
line-length = 140

[tool.poe.tasks]
black = "black src"
black-check = "black --check src"
flake8 = "flake8 --ignore=E501,W503 src"
